requirements.txt
Pillow==6.0.0
robotframework>=3.1.2